refactor: abstract custom-http-director; openai-compatible-director based on it
