+++ internal/apps/ai-proxy/filters/directors/openai-compatible-director/filter.go
@@ -17,18 +17,15 @@ package openai_compatible_director
 import (
 	"context"
 	"encoding/json"
-	"fmt"
 	"io"
 	"net/http"
 	"strings"
 
-	"github.com/erda-project/erda-infra/providers/component-protocol/utils/cputil"
-	"github.com/erda-project/erda-proto-go/apps/aiproxy/model/pb"
 	"github.com/erda-project/erda/internal/apps/ai-proxy/common/ctxhelper"
+	custom_http_director "github.com/erda-project/erda/internal/apps/ai-proxy/filters/directors/custom-http-director"
 	"github.com/erda-project/erda/internal/apps/ai-proxy/models/metadata"
 	"github.com/erda-project/erda/internal/apps/ai-proxy/models/metadata/api_style"
 	"github.com/erda-project/erda/pkg/reverseproxy"
-	"github.com/erda-project/erda/pkg/strutil"
 )
 
 const (
@@ -44,11 +41,11 @@ func init() {
 }
 
 type OpenaiCompatibleDirector struct {
-	*reverseproxy.DefaultResponseFilter
+	*custom_http_director.CustomHTTPDirector
 }
 
 func New(_ json.RawMessage) (reverseproxy.Filter, error) {
-	return &OpenaiCompatibleDirector{DefaultResponseFilter: reverseproxy.NewDefaultResponseFilter()}, nil
+	return &OpenaiCompatibleDirector{CustomHTTPDirector: custom_http_director.New()}, nil
 }
 
 func (f *OpenaiCompatibleDirector) MultiResponseWriter(ctx context.Context) []io.ReadWriter {
@@ -61,194 +58,4 @@ func (f *OpenaiCompatibleDirector) Enable(ctx context.Context, _ *http.Request) 
 	providerMeta := providerNormalMeta.MustToModelProviderMeta()
 	return providerMeta.Public.API != nil &&
 		strings.EqualFold(string(providerMeta.Public.API.APIStyle), string(api_style.APIStyleOpenAICompatible))
-}
-
-func (f *OpenaiCompatibleDirector) OnRequest(ctx context.Context, w http.ResponseWriter, infor reverseproxy.HttpInfor) (signal reverseproxy.Signal, err error) {
-	provider := ctxhelper.MustGetModelProvider(ctx)
-	providerNormalMeta := metadata.FromProtobuf(provider.Metadata)
-	providerMeta := providerNormalMeta.MustToModelProviderMeta()
-
-	// handle api style config
-	apiStyleConfig := providerMeta.Public.API.APIStyleConfig
-	if apiStyleConfig == nil {
-		return reverseproxy.Intercept, fmt.Errorf("APIStyleConfig is nil, please check the model provider metadata")
-	}
-
-	// method
-	if err := methodDirector(ctx, infor, *apiStyleConfig); err != nil {
-		return reverseproxy.Intercept, fmt.Errorf("failed to set method director, err: %v", err)
-	}
-
-	// schema
-	if err := schemaDirector(ctx, infor, *apiStyleConfig); err != nil {
-		return reverseproxy.Intercept, fmt.Errorf("failed to set schema director, err: %v", err)
-	}
-
-	// host
-	if err := hostDirector(ctx, infor, *apiStyleConfig); err != nil {
-		return reverseproxy.Intercept, fmt.Errorf("failed to set host director, err: %v", err)
-	}
-
-	// path
-	if err := pathDirector(ctx, infor, *apiStyleConfig); err != nil {
-		return reverseproxy.Intercept, fmt.Errorf("failed to set path director, err: %v", err)
-	}
-
-	// queryParams
-	if err := queryParamsDirector(ctx, infor, *apiStyleConfig); err != nil {
-		return reverseproxy.Intercept, fmt.Errorf("failed to set query params director, err: %v", err)
-	}
-
-	// headers
-	if err := headersDirector(ctx, infor, *apiStyleConfig); err != nil {
-		return reverseproxy.Intercept, fmt.Errorf("failed to set headers director, err: %v", err)
-	}
-
-	return reverseproxy.Continue, nil
-}
-
-func methodDirector(ctx context.Context, infor reverseproxy.HttpInfor, apiStyleConfig api_style.APIStyleConfig) error {
-	method := apiStyleConfig.Method
-	if method == "" {
-		method = http.MethodPost
-	}
-	method = handleJSONPathTemplate(ctx, method)
-	reverseproxy.AppendDirectors(ctx, func(req *http.Request) {
-		req.Method = method
-	})
-	return nil
-}
-
-func schemaDirector(ctx context.Context, infor reverseproxy.HttpInfor, apiStyleConfig api_style.APIStyleConfig) error {
-	schema := "https"
-	if apiStyleConfig.Scheme != "" {
-		schema = apiStyleConfig.Scheme
-	}
-	schema = handleJSONPathTemplate(ctx, schema)
-	reverseproxy.AppendDirectors(ctx, func(req *http.Request) {
-		req.URL.Scheme = schema
-	})
-	return nil
-}
-
-func hostDirector(ctx context.Context, infor reverseproxy.HttpInfor, apiStyleConfig api_style.APIStyleConfig) error {
-	host := apiStyleConfig.Host
-	if host == "" {
-		return fmt.Errorf("host is empty in APIStyleConfig")
-	}
-	host = handleJSONPathTemplate(ctx, host)
-	reverseproxy.AppendDirectors(ctx, func(req *http.Request) {
-		req.Host = host
-		req.URL.Host = host
-		req.Header.Set("Host", host)
-		req.Header.Set("X-Forwarded-Host", host)
-	})
-	return nil
-}
-
-func pathDirector(ctx context.Context, infor reverseproxy.HttpInfor, apiStyleConfig api_style.APIStyleConfig) error {
-	path := "/v1/chat/completions"
-	if apiStyleConfig.Path != "" {
-		path = apiStyleConfig.Path
-	}
-	// adjust by model type
-	model := ctxhelper.MustGetModel(ctx)
-	pathWithoutCompletionSuffix := strutil.TrimSuffixes(path, "/chat/completions", "/completions")
-	switch model.Type {
-	case pb.ModelType_embedding:
-		path = pathWithoutCompletionSuffix + "/embeddings"
-	case pb.ModelType_audio:
-		path = pathWithoutCompletionSuffix + "/audio/transcriptions"
-	case pb.ModelType_image:
-		path = pathWithoutCompletionSuffix + "/images/generations"
-	default:
-		// use completions suffix
-	}
-	// TODO custom directly setting
-	path = handleJSONPathTemplate(ctx, path)
-	reverseproxy.AppendDirectors(ctx, func(req *http.Request) {
-		req.URL.Path = path
-	})
-	return nil
-}
-
-func queryParamsDirector(ctx context.Context, infor reverseproxy.HttpInfor, apiStyleConfig api_style.APIStyleConfig) error {
-	if len(apiStyleConfig.QueryParams) == 0 {
-		return nil
-	}
-	reverseproxy.AppendDirectors(ctx, func(req *http.Request) {
-		query := req.URL.Query()
-		for key, values := range apiStyleConfig.QueryParams {
-			if len(values) == 0 {
-				continue
-			}
-			op := strings.ToLower(values[0])
-			switch op {
-			case "add":
-				for _, value := range values[1:] {
-					value = handleJSONPathTemplate(ctx, value)
-					query.Add(key, value)
-				}
-			case "set":
-				for _, value := range values[1:] {
-					value = handleJSONPathTemplate(ctx, value)
-					query.Set(key, value)
-				}
-			case "delete", "del", "remove":
-				query.Del(key)
-			default:
-				// if the operation is not recognized, we can just ignore it
-			}
-		}
-		req.URL.RawQuery = query.Encode()
-	})
-	return nil
-}
-
-func headersDirector(ctx context.Context, infor reverseproxy.HttpInfor, apiStyleConfig api_style.APIStyleConfig) error {
-	if len(apiStyleConfig.Headers) == 0 {
-		return nil
-	}
-	reverseproxy.AppendDirectors(ctx, func(req *http.Request) {
-		for key, values := range apiStyleConfig.Headers {
-			if len(values) == 0 {
-				continue
-			}
-			op := strings.ToLower(values[0])
-			switch op {
-			case "add":
-				for _, value := range values[1:] {
-					value = handleJSONPathTemplate(ctx, value)
-					req.Header.Add(key, value)
-				}
-			case "set":
-				for _, value := range values[1:] {
-					value = handleJSONPathTemplate(ctx, value)
-					req.Header.Set(key, value)
-				}
-			case "delete", "del", "remove":
-				req.Header.Del(key)
-			default:
-				// if the operation is not recognized, we can just ignore it
-			}
-		}
-	})
-	return nil
-}
-
-func handleJSONPathTemplate(ctx context.Context, s string) string {
-	parser := api_style.MustNewJSONPathParser(api_style.DefaultRegexpPattern, api_style.DefaultMultiChoiceSplitter)
-	if !parser.NeedDoReplace(s) {
-		return s
-	}
-	provider := ctxhelper.MustGetModelProvider(ctx)
-	model := ctxhelper.MustGetModel(ctx)
-	availableObjects := map[string]any{
-		"provider": provider,
-		"model":    model,
-	}
-	var jsonMap map[string]any
-	cputil.MustObjJSONTransfer(&availableObjects, &jsonMap)
-	result := parser.SearchAndReplace(s, jsonMap)
-	return result
 }
